requirements.txt
# DFS Lineup Optimizer - Dependencies

# Core Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Optimization
pulp>=2.7.0

# Excel Support
openpyxl>=3.1.0

# Advanced Table Component
streamlit-aggrid>=1.0.5

# Database (Phase 2C)
sqlalchemy>=2.0.0

# API Clients (Phase 2C)
requests>=2.31.0
python-dotenv>=1.0.0

# NFL Data (Phase 2C)
nfl-data-py>=0.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

